pyproject.toml
[project]
name = "tg-anime-bot"
version = "0.1.0"
description = "A Telegram bot for searching and watching anime episodes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=20.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "anipy-api",
]

[project.scripts]
anime-bot = "bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.uv]
package = true

[tool.uv.sources]
anipy-api = { path = "anipy-cli/api", editable = true }

[dependency-groups]
dev = [
    "pyrefly>=0.50.1",
]

